move switch chain logic to useTransaction instead of TransactionButton
